# lib/api-zod/src/generated/types/adminStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FightersRoom API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AdminStats {
  totalFighters: number;
  pendingApproval: number;
  approvedFighters: number;
  totalOpportunities: number;
  totalEvents: number;
  totalApplications: number;
  pendingApplications: number;
}