# lib/api-zod/src/generated/types/createApplicationBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FightersRoom API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateApplicationBody {
  /** @nullable */
  opportunityId?: number | null;
  /** @nullable */
  eventId?: number | null;
  /** @nullable */
  message?: string | null;
}