# lib/api-zod/src/generated/types/createFighterBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FightersRoom API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateFighterBody {
  fullName: string;
  country: string;
  city: string;
  weightClass: string;
  record: string;
  /** @nullable */
  age?: number | null;
  /** @nullable */
  height?: string | null;
  /** @nullable */
  stance?: string | null;
  /** @nullable */
  coach?: string | null;
  /** @nullable */
  manager?: string | null;
  /** @nullable */
  instagram?: string | null;
  /** @nullable */
  whatsapp?: string | null;
  email: string;
  /** @nullable */
  videoLinks?: string | null;
  /** @nullable */
  bio?: string | null;
  availableInternationally: boolean;
}